# SurfacePoint data structure encapsulations