--- a/Gems/GradientSignal/Code/Include/GradientSignal/Components/SurfaceMaskGradientComponent.h
+++ b/Gems/GradientSignal/Code/Include/GradientSignal/Components/SurfaceMaskGradientComponent.h
@@ -87,12 +87,14 @@ namespace GradientSignal
 
             points.EnumeratePoints([&result](
                 [[maybe_unused]] const AZ::Vector3& position, [[maybe_unused]] const AZ::Vector3& normal,
-                const SurfaceData::SurfaceTagWeightMap& masks) -> bool
+                const SurfaceData::SurfaceTagWeights& masks) -> bool
             {
-                for (const auto& [maskId, weight] : masks)
-                {
-                    result = AZ::GetMax(AZ::GetClamp(weight, 0.0f, 1.0f), result);
-                }
+                    masks.EnumerateWeights(
+                        [&result]([[maybe_unused]] AZ::Crc32 surfaceType, float weight) -> bool
+                        {
+                            result = AZ::GetMax(AZ::GetClamp(weight, 0.0f, 1.0f), result);
+                            return true;
+                        });
                 return true;
             });