--- a/Gems/GradientSignal/Code/Source/Components/GradientSurfaceDataComponent.cpp
+++ b/Gems/GradientSignal/Code/Source/Components/GradientSurfaceDataComponent.cpp
@@ -224,40 +224,36 @@ namespace GradientSignal
                 validShapeBounds = m_cachedShapeConstraintBounds.IsValid();
             }
 
-            const AZ::EntityId entityId = GetEntityId();
-            surfacePointList.EnumeratePoints(
-                [this, entityId, validShapeBounds, shapeConstraintBounds](SurfaceData::SurfacePoint& point) -> bool
+            surfacePointList.ModifySurfaceWeights(
+                GetEntityId(), 
+                [this, validShapeBounds, shapeConstraintBounds](SurfaceData::SurfacePoint& point)
                 {
-                    if (point.m_entityId != entityId)
+                    bool inBounds = true;
+
+                    // If we have an optional shape bounds, verify the point exists inside of it before querying the gradient value.
+                    // Otherwise, assume an unbounded surface modifier and allow *all* points through the shape check.
+                    if (validShapeBounds)
                     {
-                        bool inBounds = true;
-
-                        // If we have an optional shape bounds, verify the point exists inside of it before querying the gradient value.
-                        // Otherwise, assume an unbounded surface modifier and allow *all* points through the shape check.
-                        if (validShapeBounds)
+                        inBounds = false;
+                        if (shapeConstraintBounds.Contains(point.m_position))
                         {
-                            inBounds = false;
-                            if (shapeConstraintBounds.Contains(point.m_position))
-                            {
-                                LmbrCentral::ShapeComponentRequestsBus::EventResult(
-                                    inBounds, m_configuration.m_shapeConstraintEntityId,
-                                    &LmbrCentral::ShapeComponentRequestsBus::Events::IsPointInside, point.m_position);
-                            }
-                        }
-
-                        // If the point is within our allowed shape bounds, verify that it meets the gradient thresholds.
-                        // If so, then add the value to the surface tags.
-                        if (inBounds)
-                        {
-                            const GradientSampleParams sampleParams = { point.m_position };
-                            const float value = m_gradientSampler.GetValue(sampleParams);
-                            if (value >= m_configuration.m_thresholdMin && value <= m_configuration.m_thresholdMax)
-                            {
-                                SurfaceData::AddMaxValueForMasks(point.m_masks, m_configuration.m_modifierTags, value);
-                            }
+                            LmbrCentral::ShapeComponentRequestsBus::EventResult(
+                                inBounds, m_configuration.m_shapeConstraintEntityId,
+                                &LmbrCentral::ShapeComponentRequestsBus::Events::IsPointInside, point.m_position);
                         }
                     }
-                    return true;
+
+                    // If the point is within our allowed shape bounds, verify that it meets the gradient thresholds.
+                    // If so, then return the value to add to the surface tags.
+                    if (inBounds)
+                    {
+                        const GradientSampleParams sampleParams = { point.m_position };
+                        const float value = m_gradientSampler.GetValue(sampleParams);
+                        if (value >= m_configuration.m_thresholdMin && value <= m_configuration.m_thresholdMax)
+                        {
+                            AddMaxValueForMasks(point.m_masks, m_configuration.m_modifierTags, value);
+                        }
+                    }
                 });
         }
     }