--- a/Gems/GradientSignal/Code/Source/Components/SurfaceMaskGradientComponent.cpp
+++ b/Gems/GradientSignal/Code/Source/Components/SurfaceMaskGradientComponent.cpp
@@ -198,7 +198,7 @@ namespace GradientSignal
 
                     for (size_t index = 0; index < positions.size(); index++)
                     {
-                        points.clear();
+                        points.Clear();
                         surfaceDataRequests->GetSurfacePoints(positions[index], m_configuration.m_surfaceTagList, points);
                         outValues[index] = GetMaxSurfaceWeight(points);
                     }