… Gems/GradientSignal/Code/CMakeLists.txt
… Gems/GradientSignal/Code/Tests/GradientSignalReferencesTests.cpp
… Gems/GradientSignal/Code/Tests/GradientSignalSurfaceTests.cpp
… Gems/GradientSignal/Code/Tests/GradientSignalTestFixtures.cpp
… Gems/SurfaceData/Code/CMakeLists.txt
… Gems/SurfaceData/Code/Include/SurfaceData/Utility/SurfaceDataUtility.h
… Gems/SurfaceData/Code/Source/Components/SurfaceDataColliderComponent.cpp
… Gems/SurfaceData/Code/Source/Components/SurfaceDataColliderComponent.h
… Gems/SurfaceData/Code/Source/Components/SurfaceDataShapeComponent.cpp
… Gems/SurfaceData/Code/Source/Components/SurfaceDataShapeComponent.h
… Gems/SurfaceData/Code/Source/SurfaceDataSystemComponent.cpp
… Gems/SurfaceData/Code/Source/SurfaceDataSystemComponent.h
… Gems/SurfaceData/Code/Source/SurfaceDataTypes.cpp
… Gems/SurfaceData/Code/Tests/SurfaceDataBenchmarks.cpp
… Gems/SurfaceData/Code/Tests/SurfaceDataColliderComponentTest.cpp
… Gems/SurfaceData/Code/Tests/SurfaceDataTest.cpp
… Gems/SurfaceData/Code/surfacedata_files.cmake
… Gems/Terrain/Code/Source/Components/TerrainSurfaceDataSystemComponent.cpp
… Gems/Vegetation/Code/Include/Vegetation/Ebuses/AreaRequestBus.h
… Gems/Vegetation/Code/Include/Vegetation/InstanceData.h
… Gems/Vegetation/Code/Source/AreaSystemComponent.cpp
… Gems/Vegetation/Code/Source/Components/PositionModifierComponent.cpp
… Gems/Vegetation/Code/Source/Components/SpawnerComponent.cpp
… Gems/Vegetation/Code/Source/Components/SurfaceMaskDepthFilterComponent.cpp
… Gems/Vegetation/Code/Source/Components/SurfaceMaskFilterComponent.cpp
… Gems/Vegetation/Code/Source/Debugger/DebugComponent.cpp
… Gems/Vegetation/Code/Tests/VegetationComponentFilterTests.cpp
… Gems/Vegetation/Code/Tests/VegetationComponentModifierTests.cpp
… Gems/Vegetation/Code/Tests/VegetationMocks.h