This PR encapsulates several SurfacePoint data structures in their own classes as a step towards optimizing the SurfaceData bulk query APIs. The previous data structures were STL aliases and were used and exposed in a way that caused many small allocations and deallocations to occur. By wrapping them, it's possible to change the allocation and usage patterns to be friendlier for bulk allocations. The following changes were made:
* SurfacePointList: Changed to an opaque class where all the surface data is only accessed via APIs. The underlying data storage changed from a vector of SurfacePoint to separate vectors for each piece of data required. This will allow us to pass the position data around as a span<> for bulk APIs, and it enables us to remove SurfaceData::SurfacePoint, which was largely redundant (but not fully equivalent) with AzFramework::SurfaceData::SurfacePoint.
* SurfacePoint: Deleted. Removed all uses from within SurfacePointList, and replaced all remaining uses (mostly in test code) with AzFramework::SurfaceData::SurfacePoint.
* SurfaceTagWeightMap: Changed to an opaque class called SurfaceTagWeights where all the data is only accessed via APIs. This will make it easier to change the underlying implementation and storage during optimization passes.

Most of the changes below are simple name or API changes. The largest changes are in for() loops which are replaced with Enumerate*() calls, weight map value comparisons in test code, and the SurfaceTagWeights / SurfacePointList implementations themselves. All the unit tests still pass.

Benchmarks before change:
```
--------------------------------------------------------------------------------------------------
Benchmark                                                        Time             CPU   Iterations
--------------------------------------------------------------------------------------------------
SurfaceDataBenchmark/BM_GetSurfacePoints/1024                 1311 ms         1313 ms            1
SurfaceDataBenchmark/BM_GetSurfacePoints/2048                 5197 ms         5203 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromRegion/1024       2051 ms         2047 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromRegion/2048       8054 ms         8063 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromList/1024         1746 ms         1750 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromList/2048         7111 ms         7109 ms            1
```

Benchmarks after change:
```
--------------------------------------------------------------------------------------------------
Benchmark                                                        Time             CPU   Iterations
--------------------------------------------------------------------------------------------------
SurfaceDataBenchmark/BM_GetSurfacePoints/1024                 1221 ms         1219 ms            1
SurfaceDataBenchmark/BM_GetSurfacePoints/2048                 4912 ms         4922 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromRegion/1024       1847 ms         1844 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromRegion/2048       7319 ms         7328 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromList/1024         1807 ms         1797 ms            1
SurfaceDataBenchmark/BM_GetSurfacePointsFromList/2048         6708 ms         6719 ms            1
```